// src/api/Resources.d.ts
export interface Resources {
    /**
     * 作者
     */
    author: string;
    /**
     * 作者ID
     */
    authorId: number;
    /**
     * 内容
     */
    content: string;
    /**
     * 文章id
     */
    id: number;
    /**
     * 标题
     */
    title: string;
    /**
     * 类型
     */
    type: Type;
    /**
     * 更新时间
     */
    updateTime: string;
    /**
     * 上传时间
     */
    uploadTime: string;
}

/**
 * 类型
 */
export enum Type {
    Adminbase = "adminbase",
    Blog = "blog",
    Knowbase = "knowbase",
    News = "news",
    Notice = "notice",
}